src/requirements.txt
Flask==2.2
Werkzeug==2.2.0
cryptography==38.0.4
pyOpenSSL==23.0.0
Flask-Cors==3.0.10
Flask-Login
Flask-Migrate==4.0.1
Flask-Session==0.4.0
python-dotenv
#pysaml2
pytest==7.2.1
pyodbc==4.0.34

requests==2.27.1
psycopg2-binary
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
uwsgi==2.0.21
google-auth
cloud-sql-python-connector==1.1.0
pg8000==1.24.2
Flask-Caching
azure-core==1.26.4
azure-identity==1.12.0
azure-storage-blob==12.16.0
azure-storage-file-datalake==12.11.0
pandas
Flask-JWT-Extended==4.5.2
pyarrow==12.0.1
schedule==1.2.0
Flask-APScheduler==1.12.4
azure-storage-queue==12.6.0